fact_auth/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python
mediapipe
torch
torchvision
facenet-pytorch
scikit-learn
motor